Made the ROB faster by removing false overlap dependencies
